+++ hal/include/HAL/PDP.h
@@ -10,14 +10,14 @@
 #include <stdint.h>
 
 extern "C" {
-void HAL_InitializePDP(uint8_t module);
-double HAL_GetPDPTemperature(uint8_t module, int32_t* status);
-double HAL_GetPDPVoltage(uint8_t module, int32_t* status);
-double HAL_GetPDPChannelCurrent(uint8_t module, uint8_t channel,
-                                int32_t* status);
-double HAL_GetPDPTotalCurrent(uint8_t module, int32_t* status);
-double HAL_GetPDPTotalPower(uint8_t module, int32_t* status);
-double HAL_GetPDPTotalEnergy(uint8_t module, int32_t* status);
-void HAL_ResetPDPTotalEnergy(uint8_t module, int32_t* status);
-void HAL_ClearPDPStickyFaults(uint8_t module, int32_t* status);
+void HAL_InitializePDP(int32_t module);
+float HAL_GetPDPTemperature(int32_t module, int32_t* status);
+float HAL_GetPDPVoltage(int32_t module, int32_t* status);
+float HAL_GetPDPChannelCurrent(int32_t module, int32_t channel,
+                               int32_t* status);
+float HAL_GetPDPTotalCurrent(int32_t module, int32_t* status);
+float HAL_GetPDPTotalPower(int32_t module, int32_t* status);
+float HAL_GetPDPTotalEnergy(int32_t module, int32_t* status);
+void HAL_ResetPDPTotalEnergy(int32_t module, int32_t* status);
+void HAL_ClearPDPStickyFaults(int32_t module, int32_t* status);
 }
