+++ wpilibc/athena/include/SPI.h
@@ -57,9 +57,9 @@ class SPI : public SensorBase {
   void SetAccumulatorDeadband(int32_t deadband);
   int32_t GetAccumulatorLastValue() const;
   int64_t GetAccumulatorValue() const;
-  uint32_t GetAccumulatorCount() const;
+  int64_t GetAccumulatorCount() const;
   double GetAccumulatorAverage() const;
-  void GetAccumulatorOutput(int64_t& value, uint32_t& count) const;
+  void GetAccumulatorOutput(int64_t& value, int64_t& count) const;
 
  protected:
   uint8_t m_port;
